GeoMapping

GeoMapping

GeoMapping Julia package repo.

GeoMapping.azimuth — Method.

az = azimuth(lat1,lon1,lat2,lon2)

Compute azimuth, i.e. the angle between the line segment defined by the points (lat1,lon1) and (lat2,lon2) and the North. The units of all input and output parameters are degrees.

source

GeoMapping.distance — Method.

d = distance(lat1,lon1,lat2,lon2)

Compute the great-circle distance between the points (lat1,lon1) and (lat2,lon2). The units of all input and output parameters are degrees.

source

GeoMapping.reckon — Method.

lato,lono = reckon(lat,lon,range,azimuth)

Compute the coordinates of the end-point of a displacement on a sphere. lat,lon are the coordinates of the starting point, range is the covered distance of the displacements along a great circle and azimuth is the direction of the displacement relative to the North. The units of all input and output parameters are degrees. This function can also be used to define a spherical coordinate system with rotated poles.

source