GeoMapping

GeoMapping Julia package repo.

GeoMapping.azimuth — Method

az = azimuth(lat1,lon1,lat2,lon2)

Compute azimuth, i.e. the angle at (lat1,lon1) between the point (lat2,lon2) and the North, counted clockwise starting from the North. The units of all input and output parameters are degrees.

          North
            ↑
            | .
            |   . az
(lat1,lon1) +   .
             ╲ ↙
              ╲
               ╲
                * (lat2,lon2)

GeoMapping.distance — Method

d = distance(lat1,lon1,lat2,lon2)

Compute the great-circle distance between the points (lat1,lon1) and (lat2,lon2). The units of all input and output parameters are degrees.

\[d = \arccos( \sin\phi_1\cdot\sin\phi_2 + \cos\phi_1\cdot\cos\phi_2\cdot\cos(\Delta\lambda))\]

where $\phi_{1},\lambda_{1}$ and $\phi_{2},\lambda_{2}$ are the geographical latitude and longitude of two points 1 and 2 and $\Delta\lambda = \lambda_{2} - \lambda_{1}$.

GeoMapping.reckon — Method

lato,lono = reckon(lat,lon,range,azimuth)

Compute the coordinates of the end-point of a displacement on a sphere. lat,lon are the coordinates of the starting point, range is the covered distance of the displacements along a great circle and azimuth is the direction of the displacement relative to the North. The units of all input and output parameters are degrees. This function can also be used to define a spherical coordinate system with rotated poles.